requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
idna==3.6
numpy==1.26.4
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pycryptodome==3.20.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
six==1.16.0
statsmodels==0.14.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
websocket-client==1.7.0
